+++ shared/src/commonMain/kotlin/com/dnfapps/arrmatey/arr/api/model/MockMedia.kt
@@ -48,6 +48,7 @@ sealed class MockMedia(
     data object Radarr: MockMedia(title = "A Totally Awesome Movie")
     data object Lidarr: MockMedia(title = "A Totally Awesome Album")
     data object Readarr: MockMedia(title = "A Totally Awesome Book")
+    data object Listenarr: MockMedia(title = "A Totally Awesome Book")
 
 }
 //data class MockMedia(
